+++ rpc/src/v1/traits/ethcore.rs
@@ -67,9 +67,6 @@ pub trait Ethcore: Sized + Send + Sync + 'static {
 	/// Returns error when signer is disabled
 	fn unsigned_transactions_count(&self, _: Params) -> Result<Value, Error>;
 
-	/// Returns accounts information.
-	fn accounts_info(&self, _: Params) -> Result<Value, Error>;
-
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
@@ -89,7 +86,6 @@ pub trait Ethcore: Sized + Send + Sync + 'static {
 		delegate.add_method("ethcore_defaultExtraData", Ethcore::default_extra_data);
 		delegate.add_method("ethcore_gasPriceStatistics", Ethcore::gas_price_statistics);
 		delegate.add_method("ethcore_unsignedTransactionsCount", Ethcore::unsigned_transactions_count);
-		delegate.add_method("ethcore_accountsInfo", Ethcore::accounts_info);
 
 		delegate
 	}
